Move parse_frame_name to ddln.utils
In order to avoid cyclic imports
