+++ cvat/apps/engine/ddln/utils.py
@@ -1,0 +1,8 @@
+from pathlib import PurePath
+
+
+def parse_frame_name(path):
+    image_path = PurePath(path)
+    sequence_name = image_path.parents[2].name
+    frame_name = image_path.stem
+    return frame_name, sequence_name
